UI [#177] Implement IconButton component and refactor social action buttons for improved functionality
