+++ containers/frontend/web/src/features/social/social-dashboard.tsx
@@ -11,126 +11,10 @@ import {
   TabPanel,
   DisclosureGroup,
   DisclosureFull,
-  UserItem,
-  Icon,
-  TooltipTrigger,
-  Button,
-  type UserItemProps,
   TextField,
-  IconName,
 } from '@/components';
 
-function ButtonWithTooltip({
-  label,
-  placement = 'top',
-  onPress,
-  icon,
-  className,
-}: {
-  label: string;
-  placement?: 'top' | 'bottom' | 'left' | 'right';
-  onPress?: () => void;
-  icon: IconName;
-  className?: string;
-}) {
-  return (
-    <TooltipTrigger label={label} placement={placement}>
-      <Button
-        w="auto"
-        size="icon"
-        variant="secondary"
-        icon={<Icon name={icon} />}
-        aria-label={label}
-        className={className}
-        onPress={onPress}
-      />
-    </TooltipTrigger>
-  );
-}
-
-function UserListRequestActions({ username }: { username: string }) {
-  const tActions = useTranslations('features.social.actions');
-  return (
-    <>
-      <ButtonWithTooltip
-        label={tActions('reject')}
-        icon="close"
-        className="text-red-500 border-red-500"
-        onPress={() => console.log(username)}
-      />
-      <ButtonWithTooltip
-        label={tActions('accept')}
-        icon="check"
-        className="text-green-500 border-green-500"
-        onPress={() => console.log(username)}
-      />
-    </>
-  );
-}
-
-function UserListPendingActions({ username }: { username: string }) {
-  const tActions = useTranslations('features.social.actions');
-  return (
-    <ButtonWithTooltip
-      label={tActions('reject')}
-      icon="close"
-      className="text-red-500 border-red-500"
-      onPress={() => console.log(username)}
-    />
-  );
-}
-
-function UserOnlineActions({ username }: { username: string }) {
-  const tActions = useTranslations('features.social.actions');
-  return (
-    <>
-      <ButtonWithTooltip
-        label={tActions('chat')}
-        icon="messages"
-        onPress={() => console.log(username)}
-      />
-      <ButtonWithTooltip
-        label={tActions('inviteToGame')}
-        icon="gamepad"
-        onPress={() => console.log(username)}
-      />
-    </>
-  );
-}
-
-function UserOfflineActions({ username }: { username: string }) {
-  const tActions = useTranslations('features.social.actions');
-  return (
-    <ButtonWithTooltip
-      label={tActions('chat')}
-      icon="messages"
-      onPress={() => console.log(username)}
-    />
-  );
-}
-
-function UsersList({
-  friends,
-  type,
-}: {
-  friends: UserItemProps[];
-  type: 'request' | 'pending' | 'online' | 'offline';
-}) {
-  return friends.map(({ username, subtitle, avatarUrl, className }) => (
-    <UserItem
-      username={username}
-      subtitle={subtitle}
-      avatarUrl={avatarUrl}
-      className={className}
-      key={username}
-    >
-      {type === 'request' && <UserListRequestActions username={username} />}
-      {type === 'pending' && <UserListPendingActions username={username} />}
-      {type === 'online' && <UserOnlineActions username={username} />}
-      {type === 'offline' && <UserOfflineActions username={username} />}
-    </UserItem>
-  ));
-}
+import { UsersList } from './social-variants';
 
 function FriendsList() {
   const t = useTranslations('features.social.friends');
