+++ containers/frontend/web/src/features/social/social-variants/index.ts
@@ -1,0 +1,1 @@
+export { UsersList } from './users-list';
