+++ containers/frontend/web/src/features/social/social-variants/pending-buttons.tsx
@@ -1,0 +1,14 @@
+import { useTranslations } from 'next-intl';
+import { IconButton } from '@components';
+
+export function PendingButton({ username }: { username: string }) {
+  const tActions = useTranslations('features.social.actions');
+  return (
+    <IconButton
+      label={tActions('reject')}
+      icon="close"
+      className="text-red-500 border-red-500"
+      onPress={() => console.log(username)}
+    />
+  );
+}
